extract reconstructCrossTagNestedTerms into its own file
